refactor: 컴포넌트 파일명 폴더명과 통일
