+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
-import { styled } from "styled-components";
 import { ReactComponent as ShoppingCartIcon } from "../../assets/icon/logo.svg";
+import Styled from "./HeaderStyled";
 
 interface HeaderProps {
   children: React.ReactNode;
@@ -18,35 +18,6 @@ const Header = ({ children }: HeaderProps) => {
       <>{children}</>
     </Styled.Container>
   );
-};
-
-const Styled = {
-  Container: styled.div`
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    margin: 0;
-    padding: 18px 15%;
-
-    background: #333333;
-  `,
-
-  Logo: styled.div`
-    display: flex;
-    align-items: center;
-    gap: 12px;
-
-    background-color: transparent;
-
-    cursor: pointer;
-  `,
-
-  Title: styled.h1`
-    padding-top: 4px;
-    letter-spacing: 1px;
-    color: #fff;
-  `,
 };
 
 export default Header;
