+++ src/components/ProductCard/ProductImg/ProductImg.tsx
@@ -1,16 +1,8 @@
 import { memo, ImgHTMLAttributes } from "react";
-import styled from "styled-components";
+import Styled from "./ProductImgStyled";
 
 const ProductImg = ({ ...props }: ImgHTMLAttributes<HTMLImageElement>) => {
   return <Styled.Img {...props} />;
 };
 
-const Styled = {
-  Img: styled.img`
-    width: 282px;
-    height: 282px;
-
-    cursor: pointer;
-  `,
-};
 export default memo(ProductImg);
