+++ src/components/ProductCard/ProductInfo/ProductInfo.tsx
@@ -1,5 +1,5 @@
 import { memo } from "react";
-import styled from "styled-components";
+import Styled from "./ProductInfoStyled";
 
 interface ProductInfoProps {
   name: String;
@@ -14,27 +14,4 @@ const ProductInfo = ({ name, price }: ProductInfoProps) => {
   );
 };
 
-const Styled = {
-  ProductInfo: styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-
-    padding-left: 16px;
-  `,
-
-  ProductName: styled.span`
-    font-weight: 400;
-    font-size: 16px;
-
-    letter-spacing: 0.5px;
-  `,
-
-  ProductPrice: styled.span`
-    font-weight: 400;
-    font-size: 20px;
-
-    letter-spacing: 0.5px;
-  `,
-};
 export default memo(ProductInfo);
