+++ src/components/ProductCardList/ProductCardList.tsx
@@ -1,8 +1,8 @@
 import { useRecoilValue } from "recoil";
-import { styled } from "styled-components";
 import useLocalCart from "../../hooks/useLocalCart";
 import { fetchedProductListSelector } from "../../store/fetchSelectors";
-import ProductCard from "../ProductCard";
+import ProductCard from "../ProductCard/ProductCard";
+import Styled from "./ProductCardListStyled";
 
 const ProductCardList = () => {
   const fetchedProductList = useRecoilValue(fetchedProductListSelector);
@@ -17,11 +17,4 @@ const ProductCardList = () => {
   );
 };
 
-const Styled = {
-  Container: styled.ul`
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    grid-gap: 80px 45px;
-  `,
-};
 export default ProductCardList;
